Update - Refactor a lot of code - Optimize game speed
