+++ src/LuckyStrike/CharCreator.fxml
@@ -9,7 +9,7 @@
 
 <BorderPane fx:id="p" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="LuckyStrike.CharCreator">
    <center>
-      <TextArea fx:id="display" BorderPane.alignment="CENTER" />
+      <TextArea fx:id="display" editable="false" BorderPane.alignment="CENTER" />
    </center>
    <bottom>
       <VBox alignment="CENTER" prefHeight="86.0" prefWidth="600.0" spacing="20.0" BorderPane.alignment="CENTER">
